+++ Compiler/src/edu/mit/compilers/Main.java
@@ -11,9 +11,9 @@ import antlr.Token;
 import antlr.TokenStreamException;
 import antlr.collections.AST;
 import edu.mit.compilers.ast.myAST;
-import edu.mit.compilers.compile.Program;
 import edu.mit.compilers.ast.AstUtils;
 import edu.mit.compilers.grammar.*;
+import edu.mit.compilers.syntax.Program;
 import edu.mit.compilers.tools.CLI;
 import edu.mit.compilers.tools.Er;
 import edu.mit.compilers.tools.CLI.Action;
