+++ Compiler/src/edu/mit/compilers/compile/CompileBasicOperation.java
@@ -1,0 +1,106 @@
+package edu.mit.compilers.compile;
+
+import java.util.Collections;
+import java.util.List;
+
+import edu.mit.compilers.asm.Label;
+import edu.mit.compilers.asm.Num;
+import edu.mit.compilers.asm.Oprand;
+import edu.mit.compilers.asm.Reg;
+import edu.mit.compilers.asm.asm;
+import edu.mit.compilers.grammar.DecafParserTokenTypes;
+import edu.mit.compilers.st.ST;
+import edu.mit.compilers.syntax.Program;
+
+public class CompileBasicOperation {
+    public static void binaryAssign(ST st, String op, List<String> codes) {
+        if (!Program.shouldCompile()) {
+            return;
+        }
+        if (op.equals("=")) {
+            Reg tmpReg = st.newTmpReg();
+            Oprand rAddr = st.tmpPop();
+            Oprand lAddr = st.tmpPop();
+            if (rAddr instanceof Num) {
+                codes.add(
+                    asm.bin("movq", rAddr, lAddr)  
+                );
+            } else {
+                Collections.addAll(codes,
+                    asm.bin("movq", rAddr, tmpReg),
+                    asm.bin("movq", tmpReg, lAddr)
+                ); 
+            }
+        } else {
+            String asmOp = op.equals("+=") ? "addq" : "subq";
+            Reg tmpReg = st.newTmpReg();
+            Oprand rAddr = st.tmpPop();
+            Oprand lAddr = st.tmpPop();
+            if (rAddr instanceof Num) {
+                codes.add(
+                    asm.bin(asmOp, rAddr, lAddr)
+                );
+            } else {
+                Collections.addAll(codes,
+                    asm.bin("movq", lAddr, tmpReg),
+                    asm.bin(asmOp, rAddr, tmpReg),
+                    asm.bin("movq", tmpReg, lAddr)  
+                );
+            }
+        }
+    }
+
+    public static void unaryAssign(ST st, Integer operator, List<String> codes) {
+        if (!Program.shouldCompile()) {
+            return;
+        }
+        Oprand lAddr = st.tmpPop();
+        String op = operator == DecafParserTokenTypes.INCRE ? "addq" : "subq";
+        codes.add(
+            asm.bin(op, new Num(1L), lAddr)  
+        );
+    }
+
+    public static void relOps(ST st, List<String> codesCondition, List<String> codesIfExecution, List<String> codesElseExecution, List<String> codes) {
+        if (!Program.shouldCompile()) {
+            return;
+        }
+        Label ifExecutionEndLabel = new Label();
+        Label ifElseEndLabel = new Label();
+        Reg resultReg = st.newTmpReg();
+        Oprand elseOp = st.tmpPop();
+        Oprand ifOp = st.tmpPop();
+        Oprand conditionOp = st.tmpPop();
+        Reg condition;
+        if (conditionOp instanceof Reg) {
+            condition = (Reg)conditionOp;
+        } else {
+            condition = st.newTmpReg();
+            codesCondition.add(
+                asm.bin("movq", conditionOp, condition)    
+            );
+        }
+        Collections.addAll(codesCondition,
+            asm.bin("cmp", new Num(0L), condition.bite()),
+            asm.jmp("je", ifExecutionEndLabel)
+        );
+        Collections.addAll(codesIfExecution,
+            asm.bin("movq", ifOp, resultReg),
+            asm.jmp("jmp", ifElseEndLabel)
+        );
+        codesElseExecution.add(
+            asm.bin("movq", elseOp, resultReg)
+        );
+        codes.add(asm.cmt("ternary - start"));
+        codes.add(asm.cmt("ternary - condition"));
+        codes.addAll(codesCondition);
+        codes.add(asm.cmt("ternary - if execution"));
+        codes.addAll(codesIfExecution);
+        codes.add(asm.label(ifExecutionEndLabel));
+        codes.add(asm.cmt("ternary - else execution"));
+        codes.addAll(codesElseExecution);
+        codes.add(asm.label(ifElseEndLabel));
+        codes.add(asm.cmt("ternary - end"));
+        st.tmpPush(resultReg);
+    }
+}
