+++ Compiler/src/edu/mit/compilers/syntax/Element.java
@@ -1,4 +1,4 @@
-package edu.mit.compilers.compile;
+package edu.mit.compilers.syntax;
 
 import java.util.Collections;
 import java.util.List;
